feat: Combine the asset and media_type modules
