+++ src/process/scss.rs
@@ -2,7 +2,10 @@ use std::path::Path;
 
 use grass::{Options, from_path};
 
-use crate::process::{AssetError, ProcessesAssets, media_type::MediaType};
+use crate::process::{
+    AssetError, ProcessesAssets,
+    asset::{Asset, MediaType},
+};
 
 impl From<Box<grass::Error>> for AssetError {
     fn from(error: Box<grass::Error>) -> Self {
@@ -14,7 +17,7 @@ impl From<Box<grass::Error>> for AssetError {
 pub struct ScssProcessor {}
 
 impl ProcessesAssets for ScssProcessor {
-    fn process(&self, asset: &mut super::Asset) -> Result<(), AssetError> {
+    fn process(&self, asset: &mut Asset) -> Result<(), AssetError> {
         if *asset.media_type() != MediaType::Scss {
             tracing::debug!(
                 "skipping asset {}: not SCSS {}",
